#276 TDMA: int not large enough for default constant

+++ drv/tdma_client.h
@@ -39,9 +39,9 @@ typedef enum {
 typedef struct __attribute__((packed)) {
     uint32_t frame_duration;  ///< Duration of the entire TDMA frame [microseconds]
     uint32_t rx_start;        ///< Time between the start of the frame and when the gateway starts transmitting
-    uint16_t rx_duration;     ///< Duration the gateway will transmit messages
+    uint32_t rx_duration;     ///< Duration the gateway will transmit messages
     uint32_t tx_start;        ///< Time between the start of the frame and the start of the DotBot's alloted frame
-    uint16_t tx_duration;     ///< Duration of the DotBot's alloted frame.
+    uint32_t tx_duration;     ///< Duration of the DotBot's alloted frame.
 } tdma_client_table_t;
 
 typedef void (*tdma_client_cb_t)(uint8_t *packet, uint8_t length);  ///< Function pointer to the callback function called on packet receive
